Move MultithreadedTestRule
