+++ subprojects/core/src/test/groovy/org/gradle/process/internal/worker/DefaultWorkerProcessSpec.groovy
@@ -24,7 +24,7 @@ import org.gradle.process.internal.ExecHandleListener
 import org.gradle.process.internal.ExecHandleState
 import org.gradle.process.internal.health.memory.JvmMemoryStatus
 import org.gradle.test.fixtures.ConcurrentTestUtil
-import org.gradle.util.MultithreadedTestRule
+import org.gradle.util.internal.MultithreadedTestRule
 import org.junit.Rule
 import spock.lang.Specification
